Add c_compiler_executable to cc_toolchain
This allows configuring a compiler to be used for compiling C files.
Previously, the configured `gcc` tool would be used to specify both the
C and C++ compiler, however there are use cases where one might want to
use a compiler which does not work for both C and C++ (e.g. g++). With
this change, one may specify the `cc` tool which will be used to compile
C code if set (otherwise we fall back to using `gcc`).

This also exposes the `cc` tool through the cpp fragment, which should
allow rule creators to use the correct compiler.